+++ src/application/view/styling.hpp
@@ -2,16 +2,16 @@
 #define VIEW_STYLING_HPP_INCLUDED
 
 #include "domain/color.hpp"
-const float SPACING_SMALL = 2.0f;
-const float SPACING_MEDIUM = 4.0f;
-const float SPACING_LARGE = 8.0f;
-const float SPACING_XL = 16.0f;
-const float SPACING_XXL = 32.0f;
+const int SPACING_SMALL = 2;
+const int SPACING_MEDIUM = 4;
+const int SPACING_LARGE = 8;
+const int SPACING_XL = 16;
+const int SPACING_XXL = 32;
 
-const float BORDER_THICKNESS = 1.0f;
+const int BORDER_THICKNESS = 1;
 
-const float CORNER_RADIUS = 5.0f;
-const float SHADOW_LENGTH = 15.0f;
+const int CORNER_RADIUS = 5;
+const int SHADOW_LENGTH = 15;
 
 const domain::RGBAColor default_border_color = domain::RGBAColor{0.5, 0.5, 0.5, 1.0};
